requirements.txt
Flask
python-telegram-bot
requests
waitress
psycopg2-binary